Issue #154 Revised excelToJSON conversion code to share same rules for connectivity models and scaffolds. Added logging error that provides row and column numbers to locate fields with wrongly typed values. Removed some obsolete models from tests.
